Fix EZP-27152: show REST exceptions in Symfony profiler
Moved the REST exception listener down in the priority list. Since it sets a Response, and setting a Response on this event automatically stops propagation,
it can't happen before other listeners that do something with the exception, like the Symfony Profiler.

Since the Core Listener wraps exceptions inside Symfony ones, the exceptions it translates to have been added a mapping in the REST value object visitors so that the error code matches the actual error.